# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Tickly Ticket Management API
  version: '1.0.0'
  description: API spec (MVP) for Ticket Management System — tenant-aware, RBAC enforced via JWT claims.
servers:
  - url: http://localhost:5000
    description: Local dev
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TicketCreateRequest:
      type: object
      required: [title, tenant_id, department_id]
      properties:
        tenant_id:
          type: string
          format: uuid
        department_id:
          type: integer
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [Low, Medium, High, Critical]
    TicketResponse:
      type: object
      properties:
        id:
          type: integer
        tenant_id:
          type: string
        department_id:
          type: integer
        creatorId:
          type: string
          description: User id who created the ticket
          readOnly: true
        assignee_id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        eta:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
    TicketEvent:
      type: object
      properties:
        id:
          type: integer
        ticket_id:
          type: integer
        type:
          type: string
        actor_id:
          type: string
        payload_json:
          type: object
        ts:
          type: string
paths:
...
  /api/tickets:
    get:
      summary: List tickets (scope depends on JWT roles/claims)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
      responses:
        '200':
          description: tickets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketResponse'
    post:
      summary: Create ticket (web or inbound)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreateRequest'
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
  /api/tickets/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    get:
      summary: Get ticket by id (scoped)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ticket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '403':
          description: forbidden
    put:
      summary: Update ticket (fields allowed depend on role)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreateRequest'
      responses:
        '204':
          description: updated
...
  /api/tickets/{id}/assign:
    post:
      summary: Assign ticket to user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
      responses:
        '200':
          description: assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketEvent'
  /api/tickets/{id}/transition:
    post:
      summary: Transition ticket status
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to_status]
              properties:
                to_status:
                  type: string
                note:
                  type: string
      responses:
        '200':
          description: transitioned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketEvent'
...